fix(report): address review nits — help/slash/refresh-flag/docs consistency
Four small fixes from review:

* help.ts: --json flag is now CLI-only in the report help topic. The slash
  command branch in index.ts unconditionally calls formatReportHuman, so
  --json has no effect there. Also reword the footer paragraph: it claimed
  the footer "always" shows source/age/timestamp but error paths
  (MISSING_TICKER, EVENT_NOT_FOUND, NO_REPORT) skip the footer entirely.

* index.ts: report slash-command branch now checks parsed.parseErrors
  before kicking off the Octagon call. dispatch.ts does this for the CLI
  path; the slash path was silently swallowing typo'd flags and still
  burning a fetch (plus 3 credits on --refresh).

* README.md: --refresh flags-table row now references both analyze and
  report, matching what the per-command rows already document.

* report.ts / analyze.ts / utils/time.ts: dedupe the formatAge helper into
  src/utils/time.ts so the relative-age string is computed in one place.
  report.ts's "just now" branch carries over so sub-minute ages no longer
  render as "0m ago" in analyze either.
